# backend/openapi.yaml
openapi: 3.1.0
info:
  {
    title: Dive ID API,
    version: 1.0.0,
    description: "AI match strengths are ranking aids, not calibrated certainty.",
  }
paths:
  /health:
    get:
      responses:
        {
          "200":
            {
              description: Healthy,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          required: [status],
                          properties: { status: { const: ok } },
                        },
                    },
                },
            },
        }
  /v1/identifications/description:
    post:
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/DescriptionRequest" } },
            },
        }
      responses:
        "200":
          {
            description: Up to ten candidates (empty means no useful match),
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/IdentificationResponse" },
                  },
              },
          }
        "400": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
        "504": { $ref: "#/components/responses/Error" }
components:
  responses:
    Error:
      {
        description: Structured failure,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
  schemas:
    DescriptionRequest:
      type: object
      additionalProperties: false
      required: [requestId, description]
      properties:
        requestId: { type: string, format: uuid }
        description: { type: string, minLength: 5, maxLength: 2000 }
        context:
          {
            type: object,
            additionalProperties: false,
            properties:
              {
                region: { type: [string, "null"], maxLength: 200 },
                waterType:
                  {
                    type: [string, "null"],
                    enum: [saltwater, freshwater, brackish, unknown, null],
                  },
                approximateDepthMeters:
                  { type: [number, "null"], minimum: 0, maximum: 12000 },
                habitat: { type: [string, "null"], maxLength: 300 },
                approximateSize: { type: [string, "null"], maxLength: 100 },
                behavior: { type: [string, "null"], maxLength: 300 },
                observedAt: { type: [string, "null"], format: date-time },
              },
          }
    Match:
      type: object
      additionalProperties: false
      required:
        [
          speciesId,
          rank,
          commonName,
          scientificName,
          taxonomicResolution,
          confidenceCategory,
          matchScore,
          explanation,
          distinguishingFeatures,
          habitat,
          geographicRange,
          cautions,
        ]
      properties:
        {
          speciesId: { type: string, format: uuid },
          rank: { type: integer, minimum: 1, maximum: 10 },
          commonName: { type: string, maxLength: 120 },
          scientificName: { type: string, maxLength: 160 },
          taxonomicResolution: { enum: [species, genus, family, group] },
          confidenceCategory:
            {
              enum: [strong, good, possible, weak],
              description: Relative category; not probability.,
            },
          matchScore: { type: [number, "null"], minimum: 0, maximum: 1 },
          explanation: { type: string, maxLength: 1000 },
          distinguishingFeatures:
            { type: array, minItems: 1, maxItems: 8, items: { type: string } },
          habitat: { type: string, maxLength: 500 },
          geographicRange: { type: string, maxLength: 500 },
          cautions: { type: array, maxItems: 5, items: { type: string } },
        }
    IdentificationResponse:
      type: object
      required: [requestId, matches, disclaimer, generatedAt, modelVersion]
      properties:
        {
          requestId: { type: string, format: uuid },
          matches:
            {
              type: array,
              maxItems: 10,
              items: { $ref: "#/components/schemas/Match" },
            },
          disclaimer: { type: string },
          generatedAt: { type: string, format: date-time },
          modelVersion: { type: string },
        }
    Error:
      type: object
      required: [error]
      properties:
        {
          error:
            {
              type: object,
              required: [code, message, requestId],
              properties:
                {
                  code:
                    {
                      enum:
                        [
                          INVALID_REQUEST,
                          DESCRIPTION_TOO_SHORT,
                          DESCRIPTION_TOO_LONG,
                          RATE_LIMITED,
                          IDENTIFICATION_TIMEOUT,
                          IDENTIFICATION_UNAVAILABLE,
                          INVALID_PROVIDER_RESPONSE,
                          NO_MATCHES,
                          INTERNAL_ERROR,
                        ],
                    },
                  message: { type: string },
                  requestId: { type: string },
                },
            },
        }